+++ algonaut_transaction/src/lib.rs
@@ -4,5 +4,5 @@ pub mod builder;
 pub mod error;
 pub mod transaction;
 
-pub use builder::{Pay, Txn};
+pub use builder::{ConfigureAsset, Pay, Txn};
 pub use transaction::{Payment, SignedTransaction, Transaction, TransactionType};